requirements.txt
Faker==13.4.0
pandas==1.4.2
PyYAML==5.4.1
requests==2.27.1
ttkbootstrap==1.7.6
openpyxl==3.0.9
